♻️ Refacotr: create faker to use as mock data
