+++ src/components/common/sidebar/index.tsx
@@ -9,50 +9,19 @@ import purple from '@/public/assets/chip/ellipsePurpleLarge.svg';
 import yellow from '@/public/assets/chip/ellipseYellowLarge.svg';
 import crown from '@/public/assets/icon/crown.svg';
 import { useState } from 'react';
-import { faker } from '@faker-js/faker';
-
-// {
-//   "cursorId": 0,
-//   "totalCount": 0,
-//   "dashboards": [
-//     {
-//       "id": 0,
-//       "title": "string",
-//       "color": "string",
-//       "createdAt": "2024-04-19T05:42:59.505Z",
-//       "updatedAt": "2024-04-19T05:42:59.505Z",
-//       "createdByMe": true,
-//       "userId": 0
-//     }
-//   ]
-// }
+import { DASHBOARDS } from '@/faker';
+import Link from 'next/link';
+// import ReactPaginate from 'react-paginate';
 
 type Dashboard = {
   id: number;
   title: string;
   color: string;
+  createdAt: Date;
+  updatedAt: Date;
   createdByMe: boolean;
+  userId: number;
 };
-
-export const COLORS: string[] = ['green', 'blue', 'pink', 'purple', 'orange'];
-
-export function getRandomColor(): string {
-  const randomIndex = Math.floor(Math.random() * COLORS.length);
-  return COLORS[randomIndex];
-}
-
-export function createRandomDashboard(): Dashboard {
-  return {
-    id: faker.datatype.number(),
-    title: faker.internet.userName(),
-    color: getRandomColor(),
-    createdByMe: faker.datatype.boolean(),
-  };
-}
-
-export const DASHBOARDS: Dashboard[] = faker.helpers.multiple(createRandomDashboard, {
-  count: 15,
-});
 
 interface ColorSources {
   [key: string]: string;
@@ -105,34 +74,32 @@ const DashboardList = () => {
   return (
     <div className="w-full h-45 flex my-16 flex-col mobile:gap-30 mobile:translate-x-4">
       {DASHBOARDS.map((data: Dashboard) => (
-        // <Link href={`/dashboard/${data.id}`} key={data.id}>
-        <div
-          key={data.id}
-          className="flex my-6 rounded-4"
-          role="button"
-          tabIndex={0}
-          onClick={() => handleClick(data.id)}
-        >
-          <Image src={colorSources[data.color]} alt={data.color} width={8} height={8} />
-
-          {/* <Image src={`/public/assets/chip/${data.color}.svg`} alt={data.color} width={8} height={8} /> */}
-          <p
-            className={`text-18 ml-16 mr-6 tablet:text-16 tablet:ml-10 tablet:mr-4 mobile:hidden ${
-              data.id === selectedDashboard && ' text-violet'
-            }`}
+        <Link href={`/dashboard/${data.id}`} key={data.id}>
+          <div
+            key={data.id}
+            className="flex my-6 rounded-4"
+            role="button"
+            tabIndex={0}
+            onClick={() => handleClick(data.id)}
           >
-            {data.title}
-          </p>
-          {data.createdByMe && <Image src={crown} alt="crown" width={18} className="-translate-y-1 mobile:hidden" />}
-        </div>
-        // </Link>
+            <Image src={colorSources[data.color]} alt={data.color} width={8} height={8} />
+            <p
+              className={`text-18 ml-16 mr-6 tablet:text-16 tablet:ml-10 tablet:mr-4 mobile:hidden ${
+                data.id === selectedDashboard && ' text-violet'
+              }`}
+            >
+              {data.title}
+            </p>
+            {data.createdByMe && <Image src={crown} alt="crown" width={18} className="-translate-y-1 mobile:hidden" />}
+          </div>
+        </Link>
       ))}
       {/* <ReactPaginate
         breakLabel="..."
         nextLabel=">"
         onPageChange={handlePageClick}
-        pageRangeDisplayed={5}
-        pageCount={5}
+        pageRangeDisplayed={1}
+        pageCount={10}
         previousLabel="<"
         renderOnZeroPageCount={null}
         className="flex"
